Refactor search field into component
